+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Master
 
 * Add Google Analytics event dispatching immediately after sending events while debugging (@superarts)
+* Added support for Firebase Analytics (@BenchR267)
 
 ## Version 3.9.2
 * Fix issue in DSL where properties block was being invoked more than once (@arifken)